Remove unused functions and arguments
I used VSCode "Find all references" to inspect all the calls to all the
TrackingGraph functions and removed the following unused elements.
- `limit_to` args in nodes and edges. These now literally call
  networkx.nodes and networkx.edges and could be removed.
- `get_nodes_by_roi` function. This is probably unnecessary for metrics
  computation.
- `get_edges_with_attribute` function. This is never used, although
  `get_nodes_with_attribute` is used in the iou matcher. I suggest
refactoring the iou matcher to do the actual task more efficiently and
getting rid of both of these functions. This same part of the iou
matcher is the only place we use TrackingGraph.get_nodes_in_frame as
well.

Exceptions include:
- `get_locations` function. It was never called and is probably not
  necessary for computing metrics after matching is over. However, it
will be used in the point based matcher. We can revisit after that
matcher is implemented.
- `get_connected_components` and `get_tracklets` functions. They were
 never called but will likely be needed for metrics such as Cell Cycle
Accuracy. We can revisit after that metric is implemented.
- `get_node_attribute` and `get_edge_attribute`. These were just
  implemented and I will refactor the metrics to use them in the next
commit.