Separate setting flag on one node/edge from all nodes/edges
Having one function was clunky (had to cast ids specifically to
a list to detect one vs many) and inefficient (couldn't leverage
networkx functions for setting all attributes).

This commit also renames the functions to the format
`set_flag_on_node` or `set_flag_on_all_nodes` for maximum clarity.
This naming is clear that we are setting one flag on one or many
nodes or edges, and gets a head start on using `flag` instead of
`attribute` for our custom flags.