Use dict syntax to get node and edge flags
We decided that we did not want to assume that missing flags
are false, and instead explicitly annotate all flags on all
nodes/edges. The only additional functionality the getters
provided was to check for missing values and return False.
Since we do not want that functionality, we can revert to networkx
style access of attributes.